+++ frame/support/src/traits/tokens/misc.rs
@@ -1,0 +1,99 @@
+// This file is part of Substrate.
+
+// Copyright (C) 2019-2021 Parity Technologies (UK) Ltd.
+// SPDX-License-Identifier: Apache-2.0
+
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// 	http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//! Miscellaneous types.
+
+use codec::{Encode, Decode, FullCodec};
+use sp_core::RuntimeDebug;
+use sp_arithmetic::traits::AtLeast32BitUnsigned;
+
+/// One of a number of consequences of withdrawing a fungible from an account.
+#[derive(Copy, Clone, Eq, PartialEq)]
+pub enum WithdrawConsequence<Balance> {
+	/// Withdraw could not happen.
+	Failed,
+	/// Account balance would reduce to zero, potentially destroying it. The parameter is the
+	/// amount of balance which is destroyed.
+	ReducedToZero(Balance),
+	/// Account continued in existence.
+	Success,
+}
+
+/// Simple boolean for whether an account needs to be kept in existence.
+#[derive(Copy, Clone, Eq, PartialEq)]
+pub enum ExistenceRequirement {
+	/// Operation must not result in the account going out of existence.
+	///
+	/// Note this implies that if the account never existed in the first place, then the operation
+	/// may legitimately leave the account unchanged and still non-existent.
+	KeepAlive,
+	/// Operation may result in account going out of existence.
+	AllowDeath,
+}
+
+/// Status of funds.
+#[derive(PartialEq, Eq, Clone, Copy, Encode, Decode, RuntimeDebug)]
+pub enum BalanceStatus {
+	/// Funds are free, as corresponding to `free` item in Balances.
+	Free,
+	/// Funds are reserved, as corresponding to `reserved` item in Balances.
+	Reserved,
+}
+
+bitflags::bitflags! {
+	/// Reasons for moving funds out of an account.
+	#[derive(Encode, Decode)]
+	pub struct WithdrawReasons: i8 {
+		/// In order to pay for (system) transaction costs.
+		const TRANSACTION_PAYMENT = 0b00000001;
+		/// In order to transfer ownership.
+		const TRANSFER = 0b00000010;
+		/// In order to reserve some funds for a later return or repatriation.
+		const RESERVE = 0b00000100;
+		/// In order to pay some other (higher-level) fees.
+		const FEE = 0b00001000;
+		/// In order to tip a validator for transaction inclusion.
+		const TIP = 0b00010000;
+	}
+}
+
+impl WithdrawReasons {
+	/// Choose all variants except for `one`.
+	///
+	/// ```rust
+	/// # use frame_support::traits::WithdrawReasons;
+	/// # fn main() {
+	/// assert_eq!(
+	/// 	WithdrawReasons::FEE | WithdrawReasons::TRANSFER | WithdrawReasons::RESERVE | WithdrawReasons::TIP,
+	/// 	WithdrawReasons::except(WithdrawReasons::TRANSACTION_PAYMENT),
+	///	);
+	/// # }
+	/// ```
+	pub fn except(one: WithdrawReasons) -> WithdrawReasons {
+		let mut flags = Self::all();
+		flags.toggle(one);
+		flags
+	}
+}
+
+/// Simple amalgamation trait to collect together properties for an AssetId under one roof.
+pub trait AssetId: FullCodec + Copy + Default + Eq + PartialEq {}
+impl<T: FullCodec + Copy + Default + Eq + PartialEq> AssetId for T {}
+
+/// Simple amalgamation trait to collect together properties for a Balance under one roof.
+pub trait Balance: AtLeast32BitUnsigned + FullCodec + Copy + Default {}
+impl<T: AtLeast32BitUnsigned + FullCodec + Copy + Default> Balance for T {}
